Rewrite of distributions/priors stuff
This is not final - I am putting it in for review. 

The main idea is to have the distribution like we did in beast1, which can be used in a prior of elsewhere, say by operators.

Many of the distributions need testing - I may have introduced bugs on the way.