Unify the code/state nomenclature in DataType
Name the string input coming from an Alignment or similar
consistently *Sequence* (of type String), and distinguish
between *Encoding chain* (which maps 1:1 to the Sequence string) and the
actual *State set list* (which dissolves the ambiguities in the Encoding
chain.

Deprecate class methods that do not conform to this naming scheme, but
set them to be aliases of methods that do conform.

Clean derived classes to only add necessary additional methods.

Add unit tests for the sequence string→encoding chain and round trip
conversions.

Fix IntegerData unit test to use new methods.