User-specified tip date format in BEAUti. Issue #729.
This commit does a few things:
1. Causes the "Date" column of the tip date input editor table to
contain the raw trait values.
2. Stores these raw values in the corresponding TraitSet. This means
that the XML file contains the original trait values (eg. dates like
2014-03-13).
3. Uses a previously private method in the TraitSet class to process
the dates (this is the same method that BEAST uses) and populate the
"Height" column of the input editor table.
4. Moves the error handling (previously a System.exit(1)) out of the
TraitSet trait value conversion method to an exception handler in the
calling code.
5. Provides a combo-box for specifying the date format that is recorded
in the dateFormat input to TraitSet.