Fixed hidden input field prob reveaeled by input validation #732.
The "frequencies" input to the BinaryCovarion model is a required
input inhereted from the parent substitution model class.  This
input was apparently being set to "optional" at runtime via the
BinaryCovarion constructor.  However, because another input
"vfrequencies" specific to the BC class was assigned to the field
frequencies, hiding the field of the "frequencies" input, this
adjustment was in fact never made.

This issue didn't show up in unit testing because the unit test
used initByName() to initialize the BC instance, which until #732
did not check whether the input values satisfied the validation rules.