Fix bug where processDefaultValues read past the first '='
If traits defaults are specified by the user in the form `value="$(dateTrait.value=68BO_2011.0=2011.0,65BO_2013.0=2013.0,97SGR_2012.0=2012.0..."` BEAST would read past the first '=' to the last '='. This makes the parserDefinitions key equal "dateTrait.value=68BO_2011.0=2011.0,65BO_2013.0=2013.0,97SGR_2012.0" and value "2012.0". I fixed this by only setting k to the index of the first '='.